+++ config/environment.rb
@@ -17,7 +17,6 @@ require 'course_guide_methods'
 require 'type_verification'
 require 'tagger'
 require 'concept_coach/tag_generator'
-require 'doorkeeper_session_grant'
 
 %w(
   biglearn
@@ -32,3 +31,5 @@ COPYRIGHT_HOLDER = "Rice University"
 
 # Initialize the Rails application.
 Rails.application.initialize!
+
+require 'doorkeeper_session_grant'
